==> 00.5_soil-vector_rasterization.ipynb <==
## Soil vector layer rasterization

This notebook aims at properly rastering the surface depot polygon layer on the same grid as the DEM for use as an input later on with alpine3D.

This not as straightforward as one may think.

import geopandas as gpd
import rasterio
import numpy as np
from rasterio.features import rasterize
from shapely.geometry import Polygon

soil_vector = gpd.read_file('C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/vecteurs/a3d-depots-surface.gpkg')
soil_vector.head()

depot_names = soil_vector.depot.unique().tolist()
# Making sure Hydro is at the right spot with regards to the ALIPNE3D codes (needs to be 01)
depot_names.remove('Hydro')
depot_names.insert(1, 'Hydro')
depot_codes = [i for i in range(len(depot_names))]

mapping_dict = dict(zip(depot_names, depot_codes))

mapping_dict

soil_vector['depot_code'] = soil_vector.apply(lambda x : mapping_dict[x['depot']], axis=1)

# We need to explode soil_vector because some items are multipolygons
# rasterio's rasterize does not like multipolygons.
soil_vector_exploded = soil_vector.loc[:, ['depot', 'depot_code', 'geometry']].explode(ignore_index=True)
# Turns out also that bulding POLYGONS are actually POLYGONS Z. They need to be converted to 2d as well
# rasterio only likes polygons.
def convert_to_2d(geometry):
    if geometry.has_z:
        return Polygon([(x, y) for x, y, z in geometry.exterior.coords])
    return geometry

soil_vector_exploded['geometry'] = soil_vector_exploded['geometry'].apply(convert_to_2d)

soil_vector_exploded.to_file('C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/vecteurs/a3d-depots-surface_exploded.gpkg')

### Get DEM to copy the grid for soil depot rasterization

dem = rasterio.open('C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/ALPINE3D-PREP/DEM/DEM_5m_subgridding_cleaned.tif')
dem_array = dem.read(1)
dem_meta = dem.meta

shapes = list(soil_vector_exploded.apply(lambda x : (x['geometry'], x['depot_code']), axis=1))

rasterized_depots = rasterize(
    shapes=shapes,
    out_shape=(dem_meta['height'], dem_meta['width']),
    transform=dem.transform,
    all_touched=True,
    dtype=rasterio.uint8
)

# Save the rasterized depots
with rasterio.open('C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/ALPINE3D-PREP/soil-layers/depot_rasterized.tif', 'w', **dem_meta) as out_raster:
    out_raster.write(rasterized_depots, 1)
