==> 01_centroid_finder.ipynb <==
# Centroid finder

This notebooks finds the best coordinates for each HRDPS grid cell centroid by minimizing elevation difference between HRDPS cell native elevation and underlying CDEM pixels in a 800 meters radius around cell natural centroid.

import pandas as pd
import numpy as np
import geopandas as gpd
from geopandas.tools import sjoin
import rasterio

from rasterio.features import shapes
from shapely import wkt

### Step 0 : Define path to files and project CRS

# Path to the hrdps grid shapefile
hrdps_grid_path = 'C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/vecteurs/HRDPS/selected_pixels_32618.gpkg'
# Path to the 20 m resolution DEM to use
cdem_path = 'C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/a3d-prep/utm18n/dem/DEM-hrdps-grid-extent-5m-cleaned_32618.tif'

project_crs = 'EPSG:32618'

dem = rasterio.open(cdem_path)

### Step 1 : Compute DEM cells centroids

# 1. Polygonize CDEM raster
polygonized_pixels = []
with rasterio.open(cdem_path) as src:
    image = src.read(1) # first band
    polygonized_pixels = list(({'properties': {'elevation': v}, 'geometry': s} for i, (s, v) in enumerate(shapes(image, mask=None, transform=src.transform))))

polygonized_raster  = gpd.GeoDataFrame.from_features(polygonized_pixels)
# 2. Compute centroids and ditch the gemoetry column, we won't need it from here
polygonized_raster['centroid'] = polygonized_raster.apply(lambda x : x['geometry'].centroid, axis=1)
polygonized_raster.drop('geometry', inplace=True, axis=1)
polygonized_raster = polygonized_raster.set_geometry('centroid', crs='EPSG:32611')

# 1. Polygonize CDEM raster
polygonized_pixels = []
image = dem.read(1)
polygonized_pixels = list(({'properties': {'elevation': v}, 'geometry': s} for i, (s, v) in enumerate(shapes(image, mask=None, transform=dem.transform))))

polygonized_raster  = gpd.GeoDataFrame.from_features(polygonized_pixels)
# 2. Compute centroids and ditch the gemoetry column, we won't need it from here
polygonized_raster['centroid'] = polygonized_raster.apply(lambda x : x['geometry'].centroid, axis=1)
polygonized_raster.drop('geometry', inplace=True, axis=1)
polygonized_raster.rename(columns={'centroid': 'geometry'}, inplace=True)
polygonized_raster = polygonized_raster.set_geometry('geometry', crs=project_crs)

### Step 2 : For each HRDPS cell get underlying DEM pixels

# Read HRDPS grid and compute centroid
hrdps_grid = gpd.read_file(hrdps_grid_path)
# Add HRDPS centroid information
hrdps_grid['centroid'] = hrdps_grid.centroid

# Function to get elevation at a point
def get_elevation(point, dem):
    row, col = dem.index(point.x, point.y)
    elevation = dem.read(1)[row, col]
    return elevation

# Write the elevation of each HRDPS cell centroid in a new column
hrdps_grid['Elev'] = hrdps_grid['centroid'].apply(lambda point: get_elevation(point, dem))

# Join both spatial dataframes with spatial join. each CDEM pixel will be matched with corresponding HRDPS cell if it's in it.
cdem_hrdps = sjoin(polygonized_raster, hrdps_grid, how="inner")

### Step 3 : Compute distances between HRDPS centroid and underlying CDEM pixels

# Extraxt HRDPS centroids to set it as geometry and compute distance afterwards
hrdps_centroids = cdem_hrdps.loc[:, ['centroid']]
# Get a regular geometry attribute and set is as geometry column for Geopandas to not freak out
hrdps_centroids['geometry'] = hrdps_centroids['centroid']
hrdps_centroids.set_geometry('geometry', crs=project_crs)
# Compute distances
cdem_hrdps['distance'] = cdem_hrdps.geometry.distance(hrdps_centroids.geometry)
cdem_hrdps['elevation_difference'] = np.abs(cdem_hrdps['Elev'] - cdem_hrdps['elevation'])
# Sort geodataframe by station and distance
cdem_hrdps.sort_values(by=['name', 'distance'], ascending=True, inplace=True)

cdem_hrdps['diagonal'] = cdem_hrdps.apply(lambda x : np.sqrt(x['distance'] ** 2 + x['elevation']), axis=1)

cdem_hrdps.groupby('name')['diagonal'].min().reset_index()

# Get the indexes of the minimum values for each group
idx = cdem_hrdps.groupby('name')['diagonal'].idxmin()

# Use loc to get the rows corresponding to these indices
cdem_hrdps.loc[idx].reset_index(drop=True)

centers_df = cdem_hrdps.loc[idx].reset_index(drop=True)

centers_df

### Step 4 : Get closest candidate in elevation within the 800m radius

def coordinates_finder(group):
    '''
    Finds HRDPS best cell center in the CDEM grid. Algorithmn minimizes elevation difference in a 500m radius around the HRDPS cell centoid.
    '''
    center_group = group[group['distance'] <= 500]
    min_elevation_difference = center_group[center_group['elevation_difference'] == center_group['elevation_difference'].min()]
    min_distance = min_elevation_difference[min_elevation_difference['distance'] == min_elevation_difference['distance'].min()]
    
    return min_distance

custom_sorter = cdem_hrdps.groupby(by='name', as_index=False).apply(lambda x : coordinates_finder(x))
# For some weird reason the groupby operator is multiindexing the result. We are dropping the first dimension here which is useless.
custom_sorter.index = custom_sorter.index.droplevel(level=0)
# The apply function is a pandas generic function, and is not able to have the crs informations transit properly throught the operation
# We are going to retrieve the information from the geopandas complete dataframe using the cleaned index we just made
centers_df = cdem_hrdps.loc[custom_sorter.index]

#### Visualize resulting DataFrame

centers_df.info()

#### Save as pickle for later

# Save centers_df in pickle format
centers_df.to_pickle('C:/Users/PaulBillecocq/Documents/UdS/KRG_DATA/a3d-prep/hrdps-subgridding-prep/hrdps-dem-centroid-correspondance_32618.pkl')

centers_df
