==> 03_skyviewfactor_computation.ipynb <==
# Skyview Factor computation

In this notebook, we generate raster tiles for sky view factor and topographic correction factor as reported by this papers :
- Helbig, N., et al. “Parameterizing Surface Wind Speed over Complex Topography.” 2017

import geopandas as gpd
import numpy as np
import rasterio
from scipy import signal

dem = rasterio.open('C:/Users/PaulBillecocq/CodeRepositories/UdS_Code/violaine_GNP_subgrid/ancillary_data/cdem_dem_082N_EPSG32611_100x100_CLIPPED.tif')
hrdps_cells = gpd.read_file('C:/Users/PaulBillecocq/CodeRepositories/UdS_Code/violaine_GNP_subgrid/ancillary_data/hrdps_selected_cells.gpkg')
dem_array = dem.read(1)

## Part 1 : Sky View Factor Computation

### Mean squared slope

$\mu = \frac{\overline{[(z_{i,j} - z_{i+1,j})^2 + (z_{i,j} - z_{i,j+1})^2]}^{1/2}}{\Delta x\sqrt{2}}$

We will be using two convolution kernels to compute the two coordinates related sums. Those arrays will then be squared separately, then summed, averaged, and sqrted

def compute_mean_squared_slope(input_array, delta_x):
    '''
    Computes the mean squared slope according to Helbig, N., and Löwe, H. (2014), 
    Parameterization of the spatially averaged sky view factor in complex topography, J. Geophys. Res. Atmos., 119, 4616– 4625, doi:10.1002/2013JD020892.
    
    Parameters:
    input_array (np.array): array of elevations values on which to compute mean squared slope
    delta_x (float): cell resolution of the dem array in meters
    
    Returns:
    mu (float): mean squared slope value
    '''
    kernel_1 = np.array([
        [0, 0, 0],
        [0, 1, 0],
        [0, -1, 0]
    ])

    kernel_2 = np.array([
        [0, 0, 0],
        [0, 1, -1],
        [0, 0 , 0]
    ])

    # Compute convolution of both kernels on the dem array. Reduce their size to ensure that they are same size and that each cell is corresponding and that each cell was computed without using padded data
    member_1 = signal.convolve2d(input_array, kernel_1, mode='same')[1:,:-1]
    member_2 = signal.convolve2d(input_array, kernel_2, mode='same')[:-1,1:]
    mu = np.sqrt(np.mean(np.square(member_1) + np.square(member_2))) / (delta_x * np.sqrt(2))
    
    return mu

### Typical width of topographic features

$\xi = \frac{\sigma_{DEM}\sqrt{2}}{\mu}$

def compute_typical_width(mu, elevation_standard_deviation):
    '''
    Computes the typical width of topographic features (xi) in the provided DEM array, according to Helbig, N., and Löwe, H. (2014), 
    Parameterization of the spatially averaged sky view factor in complex topography, J. Geophys. Res. Atmos., 119, 4616– 4625, doi:10.1002/2013JD020892.
    
    Parameters:
    mu (float): mean squared slope value for the corresponding dem_array
    elevation_standard_deviation (float): standard deviation of elevation values on the desired DEM
    
    Returns:
    xi (float): typical width of topographic features in the given dem array (in meters)
    '''
    xi = elevation_standard_deviation * np.sqrt(2) / mu
    
    return xi

### Sky View Factor

$F_{sky}(L/\xi, \mu) = \Big ( 1 - \Big (1 - \frac{1}{(1 + a \mu^{b})^{c}} \Big ) e^{-d (L/\xi)^{-2}} \Big )$

with a = 3.354688, b = 1.998767, c = 0.20286, and d = 5.951

For more details, see : Helbig, N., and Löwe, H. (2014), Parameterization of the spatially averaged sky view factor in complex topography, J. Geophys. Res. Atmos., 119, 4616– 4625, doi:10.1002/2013JD020892.
    

def compute_sky_view_factor(mu, xi, L,):
    '''
    Computes the Sky View Factor as stated above
    
    Parameters:
    mu (float): mean squared slope compute for the unresolved fine-scale topography
    xi (float): typical width of topographic features in the given dem array (in meters)
    L (float): cell resolution of the coarse resolution Numerical Weather Prediction cell (in meters)
    
    Returns:
    F_sky (float): sky view factor paramter
    '''
    # Hard coded parameters, optimized by Helbig and Löwe in the 2014 paper.
    a = 3.354688
    b = 1.998767
    c = 0.20286
    d = 5.951
    
    F_sky = (1 - (1 - 1 / ((1 + a * mu ** b) ** c)) * np.exp(-d * (L / xi) ** (-2)))
    
    return F_sky

### Compute Sky View Factor
Basically, we need a value for $\mu$ and $\xi$ for each set of DEM pixels describing the geometry of each HRDPS' cell underlying topography. But there is one problem : HRDPS grid and DEM grid are not corresponding (HRDPS raw is not even on a plane ...).

The idea is to compute the centroid of each HRDPS cell polygon, get it's reference pixel in the DEM array, and extract a 2.5 km x 2.5 km square around it. Compute geometric parameters, and populate the information in each row of the HRDPS grid geodataframe.

# DEM cell resolution in meters
delta_x = 100
# HRDPS cell resolution meters
L = 2500

def sky_view_factor_processing_wrapper(dem_raster, gdf_row):
    '''
    This function is meant to be used in the apply framework of a pandas DataFrame to generate sky view factor parameters for each cell
    
    Parameters:
    dem_raster (rasterio object): rasterio object corresponding to the dem
    gdf_row (gpd row): row of the apply function
    '''
    dem_array = dem_raster.read(1)
    centroid = gdf_row['geometry'].centroid
    try:
        array_coords = dem_raster.index(centroid.x, centroid.y)
        # Build a 12 x 12 array centered around the centroid (e.g. 2.5 km x 2.5 km)
        array_subset = dem_array[(array_coords[0] - 12):(array_coords[0] + 13), (array_coords[1] - 12):(array_coords[1] + 13)]
        mu = compute_mean_squared_slope(array_subset, delta_x)
        xi = compute_typical_width(mu, np.std(dem_array))
        F_sky = compute_sky_view_factor(mu, xi, L)
        # Populate row with computed parameters
        gdf_row['mu'] = mu
        gdf_row['xi'] = xi
        gdf_row['F_sky'] = F_sky

        return gdf_row
    except:
        station = gdf_row['Station']
        print(f'Exception processing cell {station}')
        gdf_row['mu'] = np.nan
        gdf_row['xi'] = np.nan
        gdf_row['F_sky'] = np.nan
        
        return gdf_row

hrdps_cells = hrdps_cells.apply(lambda x : sky_view_factor_processing_wrapper(dem, x), axis=1)

#### Save hrdps shapefile with Sky View Factor parameters as attributes

hrdps_cells.set_geometry('geometry', crs='EPSG:32611', inplace=True)
hrdps_cells.dropna(inplace=True)
hrdps_cells.to_file('C:/Users/PaulBillecocq/CodeRepositories/UdS_Code/violaine_GNP_subgrid/ancillary_data/hrdps_grid_with_downscalingVW_topo_params.shp')

## Part 2 : VW Downscaling

### Elevation Laplacian

$\nabla^{2} z = \nabla^{2}z' \frac{\Delta_x}{4}$

$\nabla^{2} z' = \Big (z(x - \Delta x, y) + z(x + \Delta x, y) + z(x, y - \Delta x) + z(x, y + \Delta x) - 4z(x, y) \Big ) \big / \Delta x^{2}$

Again, we will be using a convolution kernel

def compute_elevation_laplacien(input_array, delta_x):
    '''
    Computes the discrete elevation Laplacian according to formulation above
    
    Parameters:
    input_array (np.array): array of elevations values
    delta_x (float): cell resolution of the dem array in meters
    
    Returns:
    nabla_z (np.array): elevation Laplacian
    '''
    kernel = np.array([
        [0, 1, 0],
        [1, -4, 1],
        [0, 1, 0]
    ])

    # Compute convolution of laplacian kernel on the dem array. Reduce its size to ensure that each cell was computed without using padded data
    nabla_z_array = signal.convolve2d(input_array, kernel, mode='valid') / (delta_x ** 2)
    nabla_array_padded = np.ones(input_array.shape) * np.nan
    nabla_array_padded[1:-1, 1:-1] = nabla_z_array
    nabla_array_padded = nabla_array_padded * delta_x / 4
    
    return nabla_array_padded

### Topographic downscaling factor

$X_{topo}^{dsc}(\nabla^{2}z, \mu) = \Big ( 1 - \frac{a' \nabla^{2}z}{1 + a' |\nabla^{2}z|^{b'}} \Big ) \frac{c'}{1 + d'\mu^{e'}}$

with a′ = 17.0393, b′ = 0.737, c′ = 1.0234, d′ = 0.3794, and e′ = 1.9821

For more details, see Helbig, N., et al. “Parameterizing Surface Wind Speed over Complex Topography.” Journal of Geophysical Research: Atmospheres, vol. 122, no. 2, John Wiley & Sons, Ltd, Jan. 2017, pp. 651–67, doi:10.1002/2016JD025593.

def compute_downscaling_factor(elevation_laplacian, mu):
    '''
    Computes downscaling according to the definition above.
    
    Parameters:
    elevation_laplacian_array (float): elevation laplacian value
    mu (float): mean squared slope value for the overlying HRDPS cell
    
    Returns:
    x_dsc_topo (float): downscaling factor for current pixel
    '''
    # Constant parameters c.f. Helbig et. al., 2017
    a_prime = 17.0393
    b_prime = 0.737
    c_prime = 1.0234
    d_prime = 0.3794
    e_prime = 1.9821
    
    x_dsc_topo = (1 - (a_prime * elevation_laplacian) / (1 + a_prime * np.abs(elevation_laplacian) ** b_prime)) * c_prime / (1 + d_prime + mu ** e_prime)
    
    return x_dsc_topo

compute_downscaling_factor_vectorized = np.vectorize(compute_downscaling_factor)

def subset_array(full_array, center_coordinates):
    '''
    Subsets the input array in 250*250 array centered on the pixel
    '''
    array_subset = full_array[(center_coordinates[0] - 12):(center_coordinates[0] + 13), (center_coordinates[1] - 12):(center_coordinates[1] + 13)]
    
    return array_subset
    
def downscaling_parameters_processing_wrapper(gdf_row, dem_raster, F_array, X_array, nabla_array):
    '''
    This function is meant to be used in the apply framework of a pandas DataFrame to generate sky view factor and downscaling parameter for each DEM pixel
    
    Parameters:
    dem_raster (rasterio object): rasterio object corresponding to the dem
    gdf_row (gpd row): row of the apply function
    '''
    dem_array = dem_raster.read(1)
    centroid = gdf_row['geometry'].centroid
    array_coords = dem_raster.index(centroid.x, centroid.y)
    dem_array_subset = subset_array(dem_array, array_coords)
    nabla_array_subset = subset_array(nabla_array, array_coords)
    
    mu = compute_mean_squared_slope(dem_array_subset, delta_x)
    xi = compute_typical_width(mu, np.std(dem_array))
    try:
        F_sky = compute_sky_view_factor(mu, xi, L)
        X_dsc_topo_subarray = compute_downscaling_factor_vectorized(nabla_array_subset, mu)
        # populate topo params array
        F_array[(array_coords[0] - 12):(array_coords[0] + 13), (array_coords[1] - 12):(array_coords[1] + 13)] = F_sky
        X_array[(array_coords[0] - 12):(array_coords[0] + 13), (array_coords[1] - 12):(array_coords[1] + 13)] = X_dsc_topo_subarray

        return (F_array, X_array)
    except:
        return (-999, -999)

hrdps_cells

F_array = np.ones(dem_array.shape) * np.nan
X_array = np.ones(dem_array.shape) * np.nan
nabla_array = compute_elevation_laplacien(dem_array, delta_x)

for i in range(len(hrdps_cells)):
    c_row = hrdps_cells.iloc[i]
    print(f'Processing VSTATION {c_row.name} at index {i}')
    F_array, X_array = downscaling_parameters_processing_wrapper(c_row, dem, F_array, X_array, nabla_array)

X_array

 with rasterio.open(
     'C:/Users/PaulBillecocq/CodeRepositories/UdS_Code/violaine_GNP_subgrid/ancillary_data/X_dsc_topo.tif',
     'w',
     driver='GTiff',
     height=X_array.shape[0],
     width=X_array.shape[1],
     count=1,
     dtype=X_array.dtype,
     crs=dem.crs,
     transform=dem.transform,
 ) as dst:
     dst.write(X_array, 1)

F_array

 with rasterio.open(
     'C:/Users/PaulBillecocq/CodeRepositories/UdS_Code/violaine_GNP_subgrid/ancillary_data/skf.tif',
     'w',
     driver='GTiff',
     height=F_array.shape[0],
     width=F_array.shape[1],
     count=1,
     dtype=F_array.dtype,
     crs=dem.crs,
     transform=dem.transform,
 ) as dst:
     dst.write(F_array, 1)
