==> 04_HRDPS_subgrid_parametrization.ipynb <==
# HRDPS subgrid parametrization

In this notebook we parametrize HRDPS forecast with the underlying CDEM 20x20m topography.
Variables TA, PSUM, and VW are parametrized.

- VW correction using sky view factor and toporaphic downscaling factor as computed in the SkyViewFactor_computation notebook. It follows the science described in Helbig, N., et al. “Parameterizing Surface Wind Speed over Complex Topography.” 2017.
- TA is corrected with an elevation bias statistics computed over 2 years of data
- PSUM is scaled with a lapse rate computed from William Durand (M.Sc. GRIMP student) snow boards measurements during season 2021-2022

import pandas as pd
import numpy as np
import rasterio
import geopandas as gpd
import os

# Get HRDPS cells information
hrdps_cells = gpd.read_file('C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/vecteurs/HRDPS/selected_pixels_32618.gpkg')

# Get topo indexes rasters information
skf_raster = rasterio.open('C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/a3d-prep/hrdps-subgridding-prep/skf_utm18n_32618.tif')
x_dsc_topo_raster = rasterio.open('C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/a3d-prep/hrdps-subgridding-prep/X_dsc_topo_utm18n_32618.tif')
cdem = rasterio.open('C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/a3d-prep/utm18n/dem/DEM-hrdps-grid-extent-5m-cleaned_32618.tif')

def sample_raster(x, y, raster):
    '''
    Retrieves point value for the specified coordinates in the given raster
    '''
    array_coords = raster.index(x, y)
    array = raster.read(1)
    point_value = array[array_coords[0], array_coords[1]]
    
    return point_value

def read_input_smet(smet_path):
    '''
    Reads the input smet path, output a dataframe
    '''
    columns = []
    with open(smet_path) as f:
        lines = f.readlines()
        columns = lines[9].split(' = ')[1].split()
    df = pd.read_csv(smet_path, skiprows=17, sep='\\s+', names=columns, index_col=0, date_format='%Y-%m-%dT%H:%M:%S')
    header = lines[:17]
    # :-1 to get rid of EOL character
    easting = float(lines[5].split(' = ')[1][:-1])
    northing = float(lines[6].split(' = ')[1][:-1])
    
    return df, header, easting, northing

def write_smet(filename, df, header):
    with open(filename, 'w') as f:
        for line in header:
            f.write(line)
        df.index = df.index.strftime('%Y-%m-%dT%H:%M:%S')
        df.to_csv(f, sep='\t', float_format='%.4f', header=False, lineterminator='\n', date_format='%Y-%m-%dT%H:%M:%S')
        
def get_elevation(file_path):
    '''Simple function to retrieve elevation value from a smet file metadata. 
    
    Parameters :
    file_path (string): Path to smet file
    
    Returns:
    elevation (float): The elevation of the station described in the smet.
    
    '''
    with open(file_path) as f:
        lines = f.readlines()
        elevation_line = lines[11]
        elevation = float(elevation_line.split(' = ')[1][:-1])
        
        return elevation
    
def precip_rate_correction(psum_value):
    if psum_value == 0:
        return psum_value
    else:
        psum = -0.905 * psum_value + 0.017
        return psum
    
def adjust_relative_humidity(row, elevation_difference):
    '''
    Correct for relative humidity following the dew point temperature correction equations written in MicroMet by Liston et al, 2006.

    Input:
    row: dataframe row, containing RH (unitless [0, 1]), and TA (K)
    station_elevation: int, Elevation of the measurement
    targetted_elevation: int, Elevation to correct for

    Output:
    rh_corrected: float, RH value scaled to the new elevation
    '''
    # constant parameters
    a = 611.21
    b = 17.502
    c = 272.55
    # vapor pressure coefficient df in km-1
    vapor_pressure_lapse_rate_dict = {
        'month': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
        'vapor_pressure': [0.41, 0.42, 0.4, 0.39, 0.38, 0.36, 0.33, 0.33, 0.36, 0.37, 0.4, 0.4]
    }
    vapor_pressure_lapse_rate_df = pd.DataFrame.from_dict(vapor_pressure_lapse_rate_dict, orient='columns')
    celsius_TA = row['TA'] - 273.15
    # computations
    saturation_vapor_pressure =  a * np.exp(b * celsius_TA / (c + celsius_TA))
    vapor_pressure = row['RH'] * saturation_vapor_pressure
    dewpoint_temp = c * np.log(vapor_pressure / a) / (b - np.log(vapor_pressure / a))
    vapor_pressure_lapse_rate = vapor_pressure_lapse_rate_df[vapor_pressure_lapse_rate_df['month'] == row.name.month]['vapor_pressure'].values[0] / 1000 # coefficient needed in m-1, hence divinding by 1000
    dewpoint_temp_lapse_rate = vapor_pressure_lapse_rate * c / b
    dewpoint_temp_corrected = dewpoint_temp - dewpoint_temp_lapse_rate * elevation_difference
    vapor_pressure_corrected = a * np.exp(b * dewpoint_temp_corrected / (c + dewpoint_temp_corrected))
    rh_corrected = vapor_pressure_corrected / saturation_vapor_pressure

    return rh_corrected

### Perform parametrization

station_list = hrdps_cells['name'].to_list()

for station in station_list:
    # Retrieve sky view foctor, topography downscaling factor
    hrdps_path = f'C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/HRDPS/02_centroid-adjusted/{station}.smet'
    hrdps_df, header, easting, northing = read_input_smet(hrdps_path)
    skf = sample_raster(easting, northing, skf_raster)
    # Get elevation data for PSUM, TA statistical correction
    x_dsc_topo = sample_raster(easting, northing, x_dsc_topo_raster)
    vstation_cdem_elevation = sample_raster(easting, northing, cdem)
    hrdps_elevation = get_elevation(hrdps_path)
    elevation_difference = vstation_cdem_elevation - hrdps_elevation
    # Compute corrected VW
    vw_corr = (hrdps_df.loc[:, ['VW']] * skf) * x_dsc_topo
    hrdps_df['VW'] = vw_corr
    # Compute corrected PSUM
    hrdps_df['PSUM'] = hrdps_df.loc[:, ['PSUM']] + 0.0011 * hrdps_df.loc[:, ['PSUM']] * elevation_difference
    # Compute correct ILWR
    delta_radiation = -2.9 * elevation_difference / 100
    # Perform lapse-rate correction on ILWR
    hrdps_df['ILWR'] = hrdps_df['ILWR'] - delta_radiation
    # Compute corrected TA
    # TA coefficient was computed using Kilpeläinen et. al., 2010
    # DOI: 10.1111/j.1600-0870.2010.00481.x
    hrdps_df['TA'] = hrdps_df.loc[:, ['TA']] + elevation_difference * 0.00628
    hrdps_df['RH'] = hrdps_df.apply(lambda x : adjust_relative_humidity(x, elevation_difference), axis=1)
    # Perform 10min interpolation
    # Resample to 10-minute intervals
    hrdps_df_resampled = hrdps_df.resample('10min').asfreq()
    
    # Interpolate all columns except PSUM
    columns_to_interpolate = ['TA', 'RH', 'VW', 'DW', 'ISWR', 'ILWR']
    hrdps_df_resampled[columns_to_interpolate] = hrdps_df_resampled[columns_to_interpolate].interpolate()
    
    # Special treatment for PSUM: equally divide each hour's value among the six 10-minute intervals
    hrdps_df_resampled['PSUM'] = hrdps_df['PSUM'].resample('10min').ffill() / 6
    # Constrain RH between 0.5 and 1
    hrdps_df_resampled.loc[hrdps_df_resampled['RH'] > 1, 'RH'] = 1
    hrdps_df_resampled.loc[hrdps_df_resampled['RH'] < 0.5, 'RH'] = 0.5
    # Write file
    dest = f'C:/Users/PaulBillecocq/OneDrive - HOROS/Documents/UdS/KRG_DATA/HRDPS/04_parameterized_psum_mm/{station}.smet'
    write_smet(dest, hrdps_df_resampled, header)

easting

station

station_list
